requirements/base.txt
alabaster==0.7.10
amqp==1.4.9
anyjson==0.3.3
argon2-cffi==16.3.0
awesome-slugify==1.6.5
Babel==2.5.1
billiard==3.3.0.23
celery==3.1.25
certifi==2017.7.27.1
cffi==1.11.0
chardet==3.0.4
coverage==4.4.1
decorator==4.1.2
defusedxml==0.5.0
Django==1.11.6
django-allauth==0.33.0
django-coverage-plugin==1.5.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-environ==0.4.4
django-extensions==1.9.1
django-model-utils==3.0.0
django-redis==4.8.0
django-test-plus==1.0.18
docutils==0.14
factory-boy==2.9.2
Faker==0.8.4
idna==2.6
imagesize==0.7.1
ipdb==0.10.3
ipython==6.2.0
ipython-genutils==0.2.0
jedi==0.11.0
Jinja2==2.9.6
kombu==3.0.37
MarkupSafe==1.0
oauthlib==2.0.4
olefile==0.44
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
py==1.4.34
pycparser==2.18
Pygments==2.2.0
pytest==3.2.2
pytest-django==3.1.2
pytest-sugar==0.9.0
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.3
redis==2.10.6
regex==2017.9.23
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
sqlparse==0.2.3
termcolor==1.1.0
traitlets==4.3.2
Unidecode==0.4.21
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2
whitenoise==3.3.0
